feat(dynamic-menu): [518] move menu logic to menu.tsx
